+++ TrackpadComponent.cpp
@@ -61,7 +61,7 @@ std::tuple<int, int> TrackpadComponent::calculatePixelPosition(float relX, float
     };
 }
 
-float TrackpadComponent::euclideanDistance(float x1, float y1, float x2, float y2) {
+double TrackpadComponent::euclideanDistance(float x1, float y1, float x2, float y2) {
     return std::sqrt(std::pow(x1 - x2, 2) + std::pow(y1 - y2, 2));
 }
